% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_boundaries.R
\name{get_year}
\alias{get_year}
\title{Get year}
\usage{
get_year(x, ...)
}
\arguments{
\item{x}{An object to be converted}

\item{...}{other parameters for methods}
}
\description{
Transforms an atomic vector to an integer year.
}
\keyword{internal}
