% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hist_town}
\alias{hist_town}
\title{County towns}
\format{
An \code{sf} object with the following columns:
\describe{
  \item{code}{County code}
  \item{town}{County town name}
  \item{from}{Start year}
  \item{tom}{End year}
  \item{geometry}{POINT geometry in SWEREF99 EPSG:3006}
}
}
\usage{
data(hist_town)
}
\description{
Swedish county towns (residensstäder), as an sf object.
}
